컴퓨트 셰이더를 통한 이펙트 생성도 적용 확인

+++ LeeEngine/LeeEngine/Source/Engine/SceneProxy/FNiagaraSceneProxy.h
@@ -98,8 +98,8 @@ struct FParticleModule
 
 struct FParticleSpawn
 {
-	int	SpawnCount;	
-	int arrPaddding[3];
+	UINT	SpawnCount;	
+	float arrPaddding[3];
 };
 
 class FNiagaraSceneProxy : public FPrimitiveSceneProxy
